3 more problems pending
